refactor(core): deduplicate helpers and drop dead exports
Hoist the duplicated metadata-comment escape/unescape pair (embed,
wiki-link) into utils/metadata-comment.ts, and the duplicated id
generator (comment, version-history) into utils/id.ts.

Delete utils/typeGuards.ts: its four guards were exported but never
used. Drop the generic groupByConsecutiveField from the public surface
(it stays internal). Replace the redundant cast inside isVizelDiagramType
with an array cast and harden VALID_DIAGRAM_TYPES via satisfies.
